Use explicit typecast of float to int.

+++ devices/common/viacuda.cpp
@@ -301,7 +301,7 @@ uint16_t ViaCuda::calc_counter_val(const uint16_t last_val, const uint64_t& last
 {
     // calculate current counter value based on elapsed time and timer frequency
     uint64_t cur_time = TimerManager::get_instance()->current_time_ns();
-    uint32_t diff = (cur_time - last_time) / this->via_clk_dur;
+    uint32_t diff = uint32_t((cur_time - last_time) / this->via_clk_dur);
     return last_val - diff;
 }
